test(pptx): replay the package fuzz corpus on stable
Select the fuzz target's mode with the low bit of the first byte instead of
sniffing the zip magic. ooxml_opc accepts zips with data prepended and a bare
end-of-central-directory record, so any payload can now reach unzip_parts.

Share the input decoding and oracle between the target and a stable test that
replays every committed seed, so a seed that panics or breaks a limit fails
cargo test.
