+++ fuzz/fuzz_targets/pptx-package-parse.rs
@@ -1,137 +1,10 @@
 #![no_main]
 
-use std::borrow::Cow;
+#[path = "../../crates/pptx-parse/tests/fuzz_harness/mod.rs"]
+mod fuzz_harness;
 
 use libfuzzer_sys::fuzz_target;
-use pptx_parse::{
-    GraphicFrameData, ParseLimits, PptxPackage, ShapeNode, TextBody, parse_pptx_with_limits,
-};
-
-const ZIP_MAGIC: &[u8] = b"PK\x03\x04";
-
-/// Just above what the richest committed fixture spends, so one mutation can
-/// cross any count limit.
-fn limits() -> ParseLimits {
-    ParseLimits {
-        max_xml_bytes: 4 << 20,
-        max_xml_events: 200_000,
-        max_xml_text_bytes: 4 << 20,
-        max_xml_depth: 64,
-        max_attributes_per_element: 128,
-        max_attribute_bytes: 64 << 10,
-        max_relationships: 64,
-        max_shapes: 96,
-        max_paragraphs: 96,
-        max_runs: 128,
-        max_comments: 16,
-    }
-}
-
-/// A raw zip as-is, or NUL-separated `path\ncontent` records rezipped so that
-/// mutated XML carries a valid CRC past `unzip_parts`.
-fn package_bytes(data: &[u8]) -> Option<Cow<'_, [u8]>> {
-    if data.starts_with(ZIP_MAGIC) {
-        return Some(Cow::Borrowed(data));
-    }
-    let parts = data
-        .split(|byte| *byte == 0)
-        .map(|record| {
-            let split = record
-                .iter()
-                .position(|byte| *byte == b'\n')
-                .unwrap_or(record.len());
-            let (path, content) = record.split_at(split);
-            (
-                String::from_utf8_lossy(path).into_owned(),
-                content.get(1..).unwrap_or_default().to_vec(),
-            )
-        })
-        .collect::<Vec<_>>();
-    ooxml_opc::rezip_parts(&parts).ok().map(Cow::Owned)
-}
-
-#[derive(Default)]
-struct Counts {
-    shapes: usize,
-    paragraphs: usize,
-    runs: usize,
-}
-
-impl Counts {
-    fn add_shapes(&mut self, shapes: &[ShapeNode]) {
-        for shape in shapes {
-            self.shapes += 1;
-            match shape {
-                ShapeNode::Shape(shape) => {
-                    if let Some(text) = &shape.text {
-                        self.add_text(text);
-                    }
-                }
-                ShapeNode::Picture(_) => {}
-                ShapeNode::GraphicFrame(frame) => match &frame.data {
-                    GraphicFrameData::Table(table) => {
-                        for cell in table.rows.iter().flat_map(|row| &row.cells) {
-                            self.add_text(&cell.text);
-                        }
-                    }
-                    GraphicFrameData::Unknown {
-                        picture: Some(_), ..
-                    } => self.shapes += 1,
-                    _ => {}
-                },
-                ShapeNode::Group(group) => self.add_shapes(&group.children),
-            }
-        }
-    }
-
-    fn add_text(&mut self, body: &TextBody) {
-        self.paragraphs += body.paragraphs.len();
-        self.runs += body
-            .paragraphs
-            .iter()
-            .map(|paragraph| paragraph.runs.len())
-            .sum::<usize>();
-    }
-}
-
-fn assert_within_limits(package: &PptxPackage, limits: &ParseLimits) {
-    let mut counts = Counts::default();
-    for slide in &package.slides {
-        counts.add_shapes(&slide.shapes);
-    }
-    for layout in &package.layouts {
-        counts.add_shapes(&layout.shapes);
-    }
-    for master in &package.masters {
-        counts.add_shapes(&master.shapes);
-    }
-    let relationships = package.relationships.values().map(Vec::len).sum::<usize>();
-    let comments = package.comments.len() + package.comment_authors.len();
-
-    assert!(
-        counts.shapes <= limits.max_shapes,
-        "shapes {}",
-        counts.shapes
-    );
-    assert!(
-        counts.paragraphs <= limits.max_paragraphs,
-        "paragraphs {}",
-        counts.paragraphs
-    );
-    assert!(counts.runs <= limits.max_runs, "runs {}", counts.runs);
-    assert!(
-        relationships <= limits.max_relationships,
-        "relationships {relationships}"
-    );
-    assert!(comments <= limits.max_comments, "comments {comments}");
-}
 
 fuzz_target!(|data: &[u8]| {
-    let Some(bytes) = package_bytes(data) else {
-        return;
-    };
-    let limits = limits();
-    if let Ok(package) = parse_pptx_with_limits(&bytes, &limits) {
-        assert_within_limits(&package, &limits);
-    }
+    let _ = fuzz_harness::run(data);
 });
